refactor: extract promo codes data logic to custom hook
